update rsyslog fixture
